@@ -1,6 +1,3 @@
-using System;
-using System.Threading;
-using System.Threading.Tasks;
 using Boilerplate.Api.Domain.Services;
 using Boilerplate.Api.Domain.Exceptions;
 using MediatR;
@@ -14,7 +11,7 @@ public static class UpdateAccountPassword
 {
     public record Command(Guid ResetPasswordToken, string Password) : IRequest;
 
-    public class Handler : AsyncRequestHandler<Command>
+    public class Handler : IRequestHandler<Command>
     {
         private readonly AppDbContext _dbContext;
         private readonly IPasswordService _passwordService;
@@ -25,7 +22,7 @@ public static class UpdateAccountPassword
             _passwordService = passwordService;
         }
 
-        protected override async Task Handle(Command request, CancellationToken cancellationToken)
+        public async Task Handle(Command request, CancellationToken cancellationToken)
         {
             var account = await _dbContext.Accounts.FirstOrDefaultAsync(x => x.ResetPasswordToken == request.ResetPasswordToken);
             if (account == null)
