@@ -1,22 +1,23 @@
-using Microsoft.EntityFrameworkCore;
-using MediatR;
-using Boilerplate.Api.Domain.Commands.Accounts;
-using Boilerplate.Api.Infrastructure.Database;
-using Boilerplate.Api.Domain.Services;
-using SendGrid;
-using Boilerplate.Api.Infrastructure.Authorization;
-using Boilerplate.Api.Infrastructure;
+
 using System.Text.Json.Serialization;
+using Boilerplate.Api;
+using Boilerplate.Api.Domain.Commands.Accounts;
+using Boilerplate.Api.Domain.PipelineBehaviours;
+using Boilerplate.Api.Domain.Services;
+using Boilerplate.Api.Infrastructure;
+using Boilerplate.Api.Infrastructure.Authorization;
+using Boilerplate.Api.Infrastructure.Database;
 using Boilerplate.Api.Infrastructure.ErrorHandling;
 using Boilerplate.Api.Infrastructure.Middleware;
-using Boilerplate.Api;
 using Boilerplate.Api.Infrastructure.Swagger;
+using MediatR;
+using Microsoft.EntityFrameworkCore;
+using SendGrid;
 using Serilog;
 using Serilog.Events;
-using Boilerplate.Api.Domain.PipelineBehaviours;
 
 var builder = WebApplication.CreateBuilder(args);
-var appsettings = builder.Configuration.Get<Appsettings>();
+var appsettings = builder.Configuration.Get<Appsettings>()!;
 
 // init serilog
 Log.Logger = new LoggerConfiguration()
@@ -36,8 +37,6 @@ builder.Services.AddControllers().AddJsonOptions(opt =>
         opt.JsonSerializerOptions.Converters.Add(new JsonStringEnumConverter());
     })
     .ConfigureApiBehaviorOptions();
-builder.Services.AddEndpointsApiExplorer();
-builder.Services.AddSwaggerGen();
 builder.Services.Configure<Appsettings>(builder.Configuration);
 builder.Services.Configure<SendGridSettings>(builder.Configuration.GetSection(nameof(Appsettings.SendGrid)));
 builder.Services.Configure<AuthorizationSettings>(builder.Configuration.GetSection(nameof(Appsettings.Authorization)));
@@ -46,13 +45,14 @@ builder.Services.AddScoped<ISendGridClientFacade, SendGridClientFacade>();
 builder.Services.AddSingleton<IHttpContextAccessor, HttpContextAccessor>();
 builder.Services.AddScoped<ISendGridClient, SendGridClient>(serviceProvider => new SendGridClient(appsettings.SendGrid.ApiKey));
 builder.Services.AddScoped<IJwtTokenHelper, JwtTokenHelper>();
-builder.Services.AddMediatR(typeof(Boilerplate.Api.Infrastructure.Database.AppDbContext).Assembly); // use any type from Boilerplate.Api
+builder.Services.AddMediatR(cfg => cfg.RegisterServicesFromAssembly(typeof(Program).Assembly));
 builder.Services.AddTransient(typeof(IPipelineBehavior<,>), typeof(LoggingBehavior<,>));
 builder.Services.AddDbContext<AppDbContext>(opts =>
 {
-    opts.UseSqlServer(builder.Configuration.GetConnectionString(nameof(ConnectionStrings.DbConnection)));
+    opts.UseMySql(
+        builder.Configuration.GetConnectionString(nameof(ConnectionStrings.DbConnection)),
+        new MySqlServerVersion(new Version(8, 0, 34))); // found in phpmyadmin by executing SELECT VERSION();
 });
-
 builder.Services.AddAuth(appsettings.Authorization);
 builder.Services.AddCustomSwagger();
 var app = builder.Build();
@@ -65,11 +65,10 @@ if (app.Environment.EnvironmentName != "Testing") // dont run this for integrati
         var dbContext = scope.ServiceProvider.GetService<AppDbContext>()!;
         var mediator = scope.ServiceProvider.GetService<IMediator>()!;
         await dbContext.Database.MigrateAsync();
-        // await mediator.Send(new EnsureDefaultAdminAccounts.Command());
+        await mediator.Send(new EnsureDefaultAdminAccounts.Command());
     }
 }
 
-// ====== Pipeline ======
 app.UseCustomSwagger();
 app.UseHsts();
 app.UseMiddleware<ExceptionMiddleware>();
